refactor legacy path
